+++ agent/agent-tooling/src/main/java/com/azure/monitor/opentelemetry/exporter/implementation/logging/AggregatingLogger.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.common;
+package com.azure.monitor.opentelemetry.exporter.implementation.logging;
 
 import java.util.Comparator;
 import java.util.HashMap;
@@ -28,6 +28,8 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+
+import com.azure.monitor.opentelemetry.exporter.implementation.utils.ThreadPoolUtils;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
