+++ agent/agent-tooling/src/main/java/com/azure/monitor/opentelemetry/exporter/implementation/pipeline/TelemetryPipelineRequest.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.telemetry;
+package com.azure.monitor.opentelemetry.exporter.implementation.pipeline;
 
 import com.azure.core.http.HttpMethod;
 import com.azure.core.http.HttpRequest;
