+++ agent/agent-tooling/src/main/java/com/azure/monitor/opentelemetry/exporter/implementation/pipeline/TelemetryPipelineResponse.java
@@ -19,52 +19,23 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.telemetry;
+package com.azure.monitor.opentelemetry.exporter.implementation.pipeline;
 
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.List;
+public class TelemetryPipelineResponse {
 
-class ByteBufferOutputStream extends OutputStream {
+  private final int statusCode;
+  private final String body;
 
-  private final AppInsightsByteBufferPool byteBufferPool;
-
-  private final List<ByteBuffer> byteBuffers = new ArrayList<>();
-
-  private ByteBuffer current;
-
-  ByteBufferOutputStream(AppInsightsByteBufferPool byteBufferPool) {
-    this.byteBufferPool = byteBufferPool;
-    current = byteBufferPool.remove();
-    byteBuffers.add(current);
+  TelemetryPipelineResponse(int statusCode, String body) {
+    this.statusCode = statusCode;
+    this.body = body;
   }
 
-  @Override
-  public void write(int b) {
-    ensureSomeCapacity();
-    current.put((byte) b);
+  public int getStatusCode() {
+    return statusCode;
   }
 
-  @Override
-  public void write(byte[] bytes, int off, int len) {
-    ensureSomeCapacity();
-    int numBytesWritten = Math.min(current.remaining(), len);
-    current.put(bytes, off, numBytesWritten);
-    if (numBytesWritten < len) {
-      write(bytes, off + numBytesWritten, len - numBytesWritten);
-    }
-  }
-
-  void ensureSomeCapacity() {
-    if (current.remaining() > 0) {
-      return;
-    }
-    current = byteBufferPool.remove();
-    byteBuffers.add(current);
-  }
-
-  List<ByteBuffer> getByteBuffers() {
-    return byteBuffers;
+  public String getBody() {
+    return body;
   }
 }
