+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/init/AppIdSupplier.java
@@ -27,8 +27,8 @@ import com.azure.core.http.HttpMethod;
 import com.azure.core.http.HttpRequest;
 import com.azure.core.http.HttpResponse;
 import com.microsoft.applicationinsights.agent.internal.common.NetworkFriendlyExceptions;
-import com.microsoft.applicationinsights.agent.internal.common.ThreadPoolUtils;
-import com.microsoft.applicationinsights.agent.internal.common.WarningLogger;
+import com.azure.monitor.opentelemetry.exporter.implementation.utils.ThreadPoolUtils;
+import com.azure.monitor.opentelemetry.exporter.implementation.logging.WarningLogger;
 import com.microsoft.applicationinsights.agent.internal.httpclient.LazyHttpClient;
 import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryClient;
 import io.opentelemetry.instrumentation.api.aisdk.AiAppId;
