+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/localstorage/LocalFileSenderTelemetryPipelineListener.java
@@ -21,9 +21,9 @@
 
 package com.microsoft.applicationinsights.agent.internal.localstorage;
 
-import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryPipelineListener;
-import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryPipelineRequest;
-import com.microsoft.applicationinsights.agent.internal.telemetry.TelemetryPipelineResponse;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineListener;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineRequest;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineResponse;
 import java.io.File;
 
 class LocalFileSenderTelemetryPipelineListener implements TelemetryPipelineListener {
