+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/telemetry/BatchItemProcessorBuilder.java
@@ -24,6 +24,7 @@ package com.microsoft.applicationinsights.agent.internal.telemetry;
 import static io.opentelemetry.api.internal.Utils.checkArgument;
 import static java.util.Objects.requireNonNull;
 
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryItemExporter;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
