+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/telemetry/DiagnosticTelemetryPipelineListener.java
@@ -21,11 +21,15 @@
 
 package com.microsoft.applicationinsights.agent.internal.telemetry;
 
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipeline;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineListener;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineRequest;
+import com.azure.monitor.opentelemetry.exporter.implementation.pipeline.TelemetryPipelineResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.microsoft.applicationinsights.agent.internal.common.NetworkFriendlyExceptions;
-import com.microsoft.applicationinsights.agent.internal.common.OperationLogger;
+import com.azure.monitor.opentelemetry.exporter.implementation.logging.OperationLogger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
